Refactor Get Involved
This page was created differently to all the other content types.

Refactoring so it is the same as everything else.
